Fixes viewer resize issue on presentation window
