+++ www/main/viewer/utils/index.js
@@ -1,0 +1,1 @@
+export { updateViewerScale } from './ScaleViewer';
